use session fixture
