+++ hq/tests/test_config.py
@@ -1,29 +1,11 @@
 # Third-party
-import pathlib
 import pytest
-import yaml
 
 # Project
 from ..config import Config
 
 
-@pytest.fixture(scope='module')
-def make_config(tmpdir_factory):
-    test_path = pathlib.Path(__file__).absolute().parent
-    config_file = (test_path / '../pkgdata' / '_test_config.yml').resolve()
-    new_config_file = tmpdir_factory.mktemp('config') / config_file.name
-    allvisit_file = test_path / 'test-allVisit.fits'
-
-    with open(config_file, 'r') as f:
-        vals = yaml.safe_load(f.read())
-    vals['input_data_file'] = str(allvisit_file)
-
-    with open(new_config_file, 'w') as f:
-        f.write(yaml.dump(vals))
-
-    return new_config_file
-
-
+@pytest.mark.usefixtures("make_config")
 def test_config(make_config):
     config_file = make_config
