+++ setup.cfg
@@ -35,7 +35,9 @@ python_requires = >=3.7
 
 [options.package_data]
 hq.pkgdata = *
-hq.tests = *.fits
+hq.tests =
+    *.fits
+    coveragerc
 
 [tool:pytest]
 testpaths = "hq" "docs"
